requirements.txt
flask===1.1.2
numpy~=1.19.2
pandas~=1.2.2
sklearn~=0.0
scikit-learn~=0.24.1
Pillow
tensorflow
